Pulling average prices of resources
Because of Albion's changes on the 24th of January, it's no longer possible to get the current in-game prices, but it's still possible to get the average prices. The calculator is now updated to give the ability to pull the average price of resources from the last 1 hour, 6 hours, 1 day, 3 days, 1 week and 1 month.